requirements.txt
pyephem==9.99
python-dotenv==1.2.1
python-telegram-bot==22.6
